+++ src/components/kytos/switch/Interface.vue
@@ -55,6 +55,30 @@ export default {
       type: Object,
       required: true,
     },
+    enabled: {
+      type: Boolean,
+      required: true,
+    },
+    active: {
+      type: Boolean,
+      required: true,
+    },
+    lldp: {
+      type: Boolean,
+      required: true,
+    },
+    nni: {
+      type: Boolean,
+      required: true,
+    },
+    uni: {
+      type: Boolean,
+      required: true,
+    },
+    metadata: {
+      type: Object,
+      required: false,
+    }
   },
   data () {
     return {